requirements.txt
langchain
openai
streamlit
python-dotenv
youtube-transcript-api
chromadb
tiktoken
faiss-cpu
wikipedia
python-dotenv
